feat(suite-native): move passphrase components
