+++ suite-native/module-passphrase/package.json
@@ -12,10 +12,14 @@
     },
     "dependencies": {
         "@react-navigation/native-stack": "7.5.1",
+        "@reduxjs/toolkit": "2.11.0",
         "@suite-common/wallet-core": "workspace:*",
         "@suite-native/device-authorization": "workspace:*",
         "@suite-native/navigation": "workspace:*",
         "@suite-native/passphrase": "workspace:*",
+        "react": "19.1.0",
+        "react-native": "0.81.5",
+        "react-native-reanimated": "~4.1.5",
         "react-redux": "9.2.0"
     }
 }
