+++ suite-native/passphrase/src/components/EmptyWalletInfoSheet.tsx
