+++ suite-native/passphrase/src/components/PassphraseScreenHeader.tsx
