+++ suite-native/passphrase/src/index.ts
@@ -1,17 +1,12 @@
 export * from './useRedirectOnPassphraseCompletion';
-export * from './PassphraseConfirmOnTrezorScreenContent';
-export * from './PassphraseScreenHeader';
-export * from './PassphraseEmptyWalletScreen';
-export * from './screens/PassphraseEnterOnTrezorScreen';
-export * from './PassphraseEnterOnTrezorScreenContent';
-export * from './PassphraseFormScreen';
-export * from './PassphraseLoadingScreen';
-export * from './PassphraseVerifyEmptyWalletScreen';
-export * from './PassphraseContentScreenWrapper';
-export * from './PassphraseForm';
-export * from './PassphraseMismatchAlert';
-export * from './PassphraseDuplicateAlert';
-export * from './PassphraseFlowDoneRedirect';
-export * from './EmptyWalletInfoSheet';
-export * from './EnterPassphraseOnTrezorButton';
-export * from './NoPassphraseButton';
+export * from './components/PassphraseConfirmOnTrezorScreenContent';
+export * from './components/PassphraseScreenHeader';
+export * from './components/PassphraseEnterOnTrezorScreenContent';
+export * from './components/PassphraseContentScreenWrapper';
+export * from './components/PassphraseForm';
+export * from './components/PassphraseMismatchAlert';
+export * from './components/PassphraseDuplicateAlert';
+export * from './components/PassphraseFlowDoneRedirect';
+export * from './components/EmptyWalletInfoSheet';
+export * from './components/EnterPassphraseOnTrezorButton';
+export * from './components/NoPassphraseButton';
